feat: Move out helpers function in a specific file
