+++ src/morphism.ts
@@ -1,70 +1,4 @@
-const aggregator = (paths: any, object: any) => {
-  return paths.reduce((delta: any, path: any) => {
-    return set(delta, path, get(object, path));
-  }, {});
-};
-
-function assignInWith(target: any, source: any, customizer?: (targetValue: any, sourceValue: any) => any) {
-  Object.entries(source).forEach(([field, value]) => {
-    if (customizer) {
-      target[field] = customizer(target[field], value);
-    } else {
-      target[field] = value;
-    }
-  });
-  return target;
-}
-
-function isUndefined(value: any) {
-  return value === undefined;
-}
-
-function isObject(value: any) {
-  const type = typeof value;
-  return value != null && (type === 'object' || type === 'function');
-}
-
-function isString(value: any): value is string {
-  return typeof value === 'string' || value instanceof String;
-}
-
-function isFunction(value: any): value is (...args: any[]) => any {
-  return typeof value === 'function';
-}
-function set(object: object, path: string, value: any) {
-  path = path.replace(/\[(\w+)\]/g, '.$1'); // convert indexes to properties
-  path = path.replace(/^\./, ''); // strip a leading dot
-  const paths = path.split('.');
-  const lastProperty = paths.pop();
-  const finalValue = paths.reduceRight(
-    (finalObject, prop) => {
-      return { [prop]: finalObject };
-    },
-    { [lastProperty]: value }
-  );
-
-  return { ...object, ...finalValue };
-}
-function get(object: object, path: string) {
-  path = path.replace(/\[(\w+)\]/g, '.$1'); // convert indexes to properties
-  path = path.replace(/^\./, ''); // strip a leading dot
-  const a = path.split('.');
-  for (let i = 0, n = a.length; i < n; ++i) {
-    const k = a[i];
-    if (isObject(object) && k in object) {
-      object = object[k];
-    } else {
-      return;
-    }
-  }
-  return object;
-}
-
-function zipObject(props: string[], values: any[]) {
-  return props.reduce((prev, prop, i) => {
-    return { ...prev, [prop]: values[i] };
-  }, {});
-}
+import { isString, get, isFunction, isObject, zipObject, isUndefined, assignInWith, aggregator } from './helpers';
 
 type ActionString = string;
 interface ActionFunction {
